Refactors out a screenwatcher object from pokefly
